+++ src/Entity/Occurrence.php
@@ -23,35 +23,36 @@ class Occurrence implements IndexItemInterface, EditableEntityInterface
     #[ORM\Id]
     #[ORM\GeneratedValue]
     #[ORM\Column]
-    #[Groups([IndexNames::Events->value])]
+    #[Groups([IndexNames::Events->value, IndexNames::Occurrences->value])]
     #[SerializedPath('[entityId]')]
     private ?int $id = null;
 
     #[ORM\Column]
-    #[Groups([IndexNames::Events->value])]
+    #[Groups([IndexNames::Events->value, IndexNames::Occurrences->value])]
     private ?\DateTimeImmutable $start = null;
 
     #[ORM\Column]
-    #[Groups([IndexNames::Events->value])]
+    #[Groups([IndexNames::Events->value, IndexNames::Occurrences->value])]
     private ?\DateTimeImmutable $end = null;
 
     #[ORM\Column(length: 255, nullable: true)]
-    #[Groups([IndexNames::Events->value])]
+    #[Groups([IndexNames::Events->value, IndexNames::Occurrences->value])]
     private ?string $ticketPriceRange = null;
 
     #[ORM\Column(length: 255, nullable: true)]
-    #[Groups([IndexNames::Events->value])]
+    #[Groups([IndexNames::Events->value, IndexNames::Occurrences->value])]
     private ?string $room = null;
 
     #[ORM\ManyToOne(inversedBy: 'occurrences')]
     #[ORM\JoinColumn(nullable: false)]
+    #[Groups([IndexNames::Occurrences->value])]
     private ?Event $event = null;
 
     #[ORM\OneToMany(mappedBy: 'occurrence', targetEntity: DailyOccurrence::class, orphanRemoval: true)]
     private Collection $dailyOccurrences;
 
     #[ORM\Column(length: 255, nullable: true)]
-    #[Groups([IndexNames::Events->value])]
+    #[Groups([IndexNames::Events->value, IndexNames::Occurrences->value])]
     private ?string $status = null;
 
     public function __construct()