+++ src/Entity/Vocabulary.php
@@ -2,15 +2,18 @@
 
 namespace App\Entity;
 
+use App\Model\Indexing\IndexNames;
 use App\Repository\VocabularyRepository;
+use App\Service\Indexing\IndexItemInterface;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\SoftDeleteable\Traits\SoftDeleteableEntity;
 use Gedmo\Timestampable\Traits\TimestampableEntity;
+use Symfony\Component\Serializer\Annotation\Groups;
 
 #[ORM\Entity(repositoryClass: VocabularyRepository::class)]
-class Vocabulary
+class Vocabulary implements IndexItemInterface
 {
     use TimestampableEntity;
     use SoftDeleteableEntity;
@@ -20,13 +23,16 @@ class Vocabulary
     #[ORM\Column]
     private ?int $id = null;
 
-    #[ORM\Column(length: 255)]
+    #[ORM\Column(length: 255, unique: true)]
+    #[Groups([IndexNames::Vocabularies->value, IndexNames::Tags->value])]
     private ?string $name = null;
 
     #[ORM\ManyToMany(targetEntity: Tag::class, inversedBy: 'vocabularies')]
+    #[Groups([IndexNames::Vocabularies->value])]
     private Collection $tags;
 
     #[ORM\Column(length: 255, nullable: true)]
+    #[Groups([IndexNames::Vocabularies->value])]
     private ?string $description = null;
 
     public function __construct()