+++ src/Repository/EventRepository.php
@@ -4,14 +4,11 @@ namespace App\Repository;
 
 use App\Entity\Event;
 use App\Model\Indexing\IndexNames;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\ORM\NonUniqueResultException;
-use Doctrine\ORM\NoResultException;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\DependencyInjection\Attribute\AsTaggedItem;
 
 /**
- * @extends ServiceEntityRepository<Event>
+ * @extends AbstractPopulateRepository<Event>
  *
  * @method Event|null find($id, $lockMode = null, $lockVersion = null)
  * @method Event|null findOneBy(array $criteria, array $orderBy = null)
@@ -19,7 +16,7 @@ use Symfony\Component\DependencyInjection\Attribute\AsTaggedItem;
  * @method Event[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */
 #[AsTaggedItem(index: IndexNames::Events->value, priority: 10)]
-final class EventRepository extends ServiceEntityRepository implements PopulateInterface
+final class EventRepository extends AbstractPopulateRepository
 {
     public function __construct(ManagerRegistry $registry)
     {
@@ -42,18 +39,5 @@ final class EventRepository extends ServiceEntityRepository implements PopulateI
         if ($flush) {
             $this->getEntityManager()->flush();
         }
-    }
-
-    /**
-     * @throws NonUniqueResultException
-     * @throws NoResultException
-     */
-    public function getNumberOfRecords(): int
-    {
-        $query = $this->createQueryBuilder('e')
-            ->select('COUNT(e.id)')
-            ->getQuery();
-
-        return $query->getSingleScalarResult();
     }
 }