+++ src/Repository/OrganizationRepository.php
@@ -4,22 +4,19 @@ namespace App\Repository;
 
 use App\Entity\Organization;
 use App\Model\Indexing\IndexNames;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\ORM\NonUniqueResultException;
-use Doctrine\ORM\NoResultException;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\DependencyInjection\Attribute\AsTaggedItem;
 
 /**
- * @extends ServiceEntityRepository<Organization>
+ * @extends AbstractPopulateRepository<Organization>
  *
  * @method Organization|null find($id, $lockMode = null, $lockVersion = null)
  * @method Organization|null findOneBy(array $criteria, array $orderBy = null)
  * @method Organization[]    findAll()
  * @method Organization[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */
-#[AsTaggedItem(index: IndexNames::Organization->value, priority: 10)]
-final class OrganizationRepository extends ServiceEntityRepository implements PopulateInterface
+#[AsTaggedItem(index: IndexNames::Organizations->value, priority: 10)]
+final class OrganizationRepository extends AbstractPopulateRepository
 {
     public function __construct(ManagerRegistry $registry)
     {
@@ -42,18 +39,5 @@ final class OrganizationRepository extends ServiceEntityRepository implements Po
         if ($flush) {
             $this->getEntityManager()->flush();
         }
-    }
-
-    /**
-     * @throws NonUniqueResultException
-     * @throws NoResultException
-     */
-    public function getNumberOfRecords(): int
-    {
-        $query = $this->createQueryBuilder('o')
-            ->select('COUNT(o.id)')
-            ->getQuery();
-
-        return $query->getSingleScalarResult();
     }
 }