+++ src/Service/Indexing/IndexingOccurrences.php
@@ -4,22 +4,20 @@ namespace App\Service\Indexing;
 
 use App\Model\Indexing\IndexFieldTypes;
 use App\Model\Indexing\IndexNames;
-use App\Service\ImageHandlerInterface;
 use Elastic\Elasticsearch\Client;
 use Symfony\Component\DependencyInjection\Attribute\AsTaggedItem;
 use Symfony\Component\Serializer\Context\Normalizer\DateTimeNormalizerContextBuilder;
 use Symfony\Component\Serializer\Context\Normalizer\ObjectNormalizerContextBuilder;
 use Symfony\Component\Serializer\SerializerInterface;
 
-#[AsTaggedItem(index: IndexNames::Events->value, priority: 10)]
-final class IndexingEvents extends AbstractIndexingElastic
+#[AsTaggedItem(index: IndexNames::Occurrences->value, priority: 10)]
+final class IndexingOccurrences extends AbstractIndexingElastic
 {
-    protected const string INDEX_ALIAS = IndexNames::Events->value;
+    protected const string INDEX_ALIAS = IndexNames::Occurrences->value;
 
     public function __construct(
-        private readonly IndexingLocations $indexingLocations,
+        private readonly IndexingEvents $indexingEvents,
         private readonly SerializerInterface $serializer,
-        private readonly ImageHandlerInterface $imageHandler,
         private readonly Client $client,
     ) {
         parent::__construct($this->client);
@@ -28,22 +26,15 @@ final class IndexingEvents extends AbstractIndexingElastic
     public function serialize(IndexItemInterface $item): array
     {
         $contextBuilder = (new ObjectNormalizerContextBuilder())
-            ->withGroups([IndexNames::Events->value]);
+            ->withGroups([IndexNames::Occurrences->value]);
         $contextBuilder = (new DateTimeNormalizerContextBuilder())
             ->withContext($contextBuilder)
             ->withFormat(IndexFieldTypes::DATEFORMAT);
+
         $data = $this->serializer->normalize($item, null, $contextBuilder->toArray());
 
         // @todo: Figure out how to do these changes with the serializer. This is just....
-        // Get tag names.
-        $data['tags'] = array_column($data['tags'], 'name');
-
-        // Fix image urls (with a full path and derived sizes).
-        $imageUrl = $data['imageUrls']['original'];
-        $data['imageUrls'] = $this->imageHandler->getTransformedImageUrls($imageUrl);
-
-        // @todo: Figure out how to do these changes with the serializer. This is just....
-        $data['location'] = $this->indexingLocations->serialize($item->getLocation());
+        $data['event'] = $this->indexingEvents->serialize($item->getEvent());
 
         return $data;
     }